refactored raw fiber photometry into its own file
